Mode 2 implemented
Mode 2 fully functional. It implements a game mode with a timer. The time limit of a game is 90 seconds. Every 10 seconds, the timer decrements. If the time is over the player is shown a score equal to the number of matches he has managed to make. If the player finishes before the timer ends, the score is computed as (time remaining/10 +nb of matches). After being shown the score, the player is taken back to the main menu to select a new mode.